requirements.txt
flask
torch
pillow
git+https://github.com/openai/CLIP.git
openai
